Moved stuff to external_class
